FIX: Ensemble class OK for DA
This continues issue #18.
Edits made to the Ensemble class to ensure it works well for data assimilation tasks. This includes:
- Adding assimilated_data attribute to store observations and times.
- Modifying update history to handle Analysis

Also, minor edits to plotting functions for better visualization of DA results.

RAISES ISSUE: DiscreteIntegrator class is not working properly. When integrating models with upsampling, the interpolator doies not function as expected. Need to investigate and fix this.